Applied changes from review
