+++ src/operator/nn/dnnl/dnnl_sum.cc
@@ -61,24 +61,7 @@ void DNNLSum(const dnnl::memory& arr1, const dnnl::memory& arr2, const dnnl::mem
   DNNLStream::Get()->RegisterPrimArgs(dnnl::sum(sum_pd), args);
 }
 
-class DNNLSumFwd {
- public:
-  dnnl::sum::primitive_desc fwd_pd;
-
-  DNNLSumFwd(const std::vector<float>& scales, const std::vector<dnnl::memory::desc>& data_md)
-      : fwd_pd(scales, data_md, CpuEngine::Get()->get_engine()) {
-    fwd_ = std::make_shared<dnnl::sum>(fwd_pd);
-  }
-
-  const dnnl::sum& GetFwd() const {
-    return *fwd_;
-  }
-
- private:
-  std::shared_ptr<dnnl::sum> fwd_;
-};
-
-static DNNLSumFwd& GetSumForward(const std::vector<float>& scales,
+DNNLSumFwd& GetSumForward(const std::vector<float>& scales,
                                  const std::vector<NDArray>& in_data,
                                  const std::vector<dnnl::memory::desc>& data_md) {
 #if DMLC_CXX11_THREAD_LOCAL
